portfolio/AI/assistants/python-fullstack-dev/api-clients/python/requirements.txt
# Core AI providers
anthropic>=0.25.0
openai>=1.12.0

# Environment and configuration
python-dotenv>=1.0.0

# Utilities
requests>=2.28.0
pathlib2>=2.3.7; python_version < '3.4'

# Optional dependencies for enhanced functionality
pydantic>=2.0.0
pyyaml>=6.0
